+++ packages/eslint-plugin-template/tests/rules/elements-content/spec.ts
@@ -1,0 +1,13 @@
+import { RuleTester } from '@angular-eslint/utils';
+import rule, { RULE_NAME } from '../../../src/rules/elements-content';
+
+import { invalid, valid } from './cases';
+
+const ruleTester = new RuleTester({
+  parser: '@angular-eslint/template-parser',
+});
+
+ruleTester.run(RULE_NAME, rule, {
+  valid,
+  invalid,
+});