+++ packages/integration-tests/utils/run-lint.ts
@@ -2,13 +2,12 @@
 import execa from 'execa';
 import path from 'path';
 
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const stripAnsi = require('strip-ansi');
+import stripAnsi = require('strip-ansi');
 
 const FIXTURES_DIR = path.join(__dirname, '../fixtures/');
 
-function normalizeOutput(value: string): string {
-  const ansiRemoved: string = stripAnsi(value);
+function normalizeOutput(value: unknown): string {
+  const ansiRemoved = stripAnsi(String(value));
   return (
     ansiRemoved
       .replace(