+++ packages/schematics/src/convert-tslint-to-eslint/schema.d.ts
@@ -1,6 +1,0 @@
-export interface Schema {
-  project?: string;
-  convertIndentationRules: boolean;
-  removeTslintIfNoMoreTslintTargets?: boolean;
-  ignoreExistingTslintConfig: boolean;
-}